Include "RetrieveUnderlying<T?>" default method in IDataProvider<T>
This should simplify the way you access the data provied by the
class, as you do not need to worry about handling the IDisposables
created by the implementations (mainly Stream<T>) anymore.

However copying the stream (which was used to backup the HTTPStreams)
will be temporarily disabled by this change.